# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: wearable
    description: Wearable device OAuth and data proxy
  - name: subscription
    description: Stripe subscription management
  - name: sync
    description: Encrypted cloud sync for Pro users
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /wearable/connect:
    get:
      operationId: wearableConnect
      tags: [wearable]
      summary: Begin OAuth flow
      description: Redirects browser to wearable provider OAuth authorization page
      responses:
        "302":
          description: Redirect to provider OAuth page

  /wearable/callback:
    get:
      operationId: wearableCallback
      tags: [wearable]
      summary: OAuth callback
      description: Handles OAuth callback, exchanges code for access token
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: error
          in: query
          required: false
          schema:
            type: string
      responses:
        "302":
          description: Redirect back to app with connection status

  /wearable/status:
    get:
      operationId: wearableStatus
      tags: [wearable]
      summary: Get connection status
      description: Returns whether a wearable is connected and provider info
      responses:
        "200":
          description: Connection status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WearableStatus"

  /wearable/disconnect:
    post:
      operationId: wearableDisconnect
      tags: [wearable]
      summary: Disconnect wearable
      description: Clears stored OAuth tokens
      responses:
        "200":
          description: Disconnected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WearableStatus"

  /wearable/data:
    get:
      operationId: wearableData
      tags: [wearable]
      summary: Fetch biometric data
      description: Proxies data fetch from wearable provider API
      parameters:
        - name: metric
          in: query
          required: true
          schema:
            type: string
            enum: [hrv, recovery, rhr, sleep]
        - name: days
          in: query
          required: false
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Biometric data points
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BiometricDataResponse"
        "401":
          description: No wearable connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /subscription/status:
    get:
      operationId: getSubscriptionStatus
      tags: [subscription]
      summary: Get subscription tier
      description: Returns the authenticated user's subscription tier (free or pro)
      security:
        - ClerkAuth: []
      responses:
        "200":
          description: Subscription status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionStatus"

  /subscription/checkout:
    post:
      operationId: createCheckoutSession
      tags: [subscription]
      summary: Create Stripe Checkout session
      description: Creates a Stripe Checkout session for upgrading to Pro. priceId is optional — backend resolves from catalog if omitted.
      security:
        - ClerkAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutRequest"
      responses:
        "200":
          description: Checkout session URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
        "503":
          description: Payments not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /subscription/portal:
    post:
      operationId: createPortalSession
      tags: [subscription]
      summary: Create Stripe customer portal session
      description: Returns a URL to the Stripe billing portal for managing the subscription
      security:
        - ClerkAuth: []
      responses:
        "200":
          description: Portal session URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortalResponse"

  /subscription/price:
    get:
      operationId: getProPrice
      tags: [subscription]
      summary: Get Protocol Pro price info
      description: Returns the active Stripe price ID and amount for Protocol Pro
      responses:
        "200":
          description: Price info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceResponse"

  /sync/salt:
    get:
      operationId: getSyncSalt
      tags: [sync]
      summary: Get per-user encryption salt
      description: Returns (or creates) a random server-side salt used in client-side key derivation for cloud sync
      security:
        - ClerkAuth: []
      responses:
        "200":
          description: Per-user salt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaltResponse"

  /sync/blob:
    get:
      operationId: downloadBlob
      tags: [sync]
      summary: Download encrypted sync blob (Pro only)
      description: Downloads the user's encrypted sync blob from the server. Requires Pro subscription.
      security:
        - ClerkAuth: []
      responses:
        "200":
          description: Encrypted blob
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlobResponse"
        "403":
          description: Pro subscription required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: uploadBlob
      tags: [sync]
      summary: Upload encrypted sync blob (Pro only)
      description: Uploads the user's AES-GCM encrypted sync blob. Requires Pro subscription.
      security:
        - ClerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlobUploadRequest"
      responses:
        "200":
          description: Upload confirmation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlobUploadResponse"
        "403":
          description: Pro subscription required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: Payload too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    ClerkAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk session token passed via Authorization header or session cookie
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    WearableStatus:
      type: object
      properties:
        connected:
          type: boolean
        provider:
          type: string
          nullable: true
        connectedAt:
          type: string
          nullable: true
      required:
        - connected
        - provider
        - connectedAt

    BiometricDataPoint:
      type: object
      properties:
        date:
          type: string
          description: ISO date string
        value:
          type: number
          nullable: true
        metric:
          type: string
      required:
        - date
        - value
        - metric

    BiometricDataResponse:
      type: object
      properties:
        metric:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/BiometricDataPoint"
        source:
          type: string
          description: "demo | whoop"
      required:
        - metric
        - data
        - source

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    SubscriptionStatus:
      type: object
      properties:
        tier:
          type: string
          enum: [free, pro]
      required:
        - tier

    CheckoutRequest:
      type: object
      properties:
        priceId:
          type: string
          description: Optional Stripe price ID; backend resolves if omitted

    CheckoutResponse:
      type: object
      properties:
        url:
          type: string
          description: Stripe Checkout session URL to redirect to
      required:
        - url

    PortalResponse:
      type: object
      properties:
        url:
          type: string
          description: Stripe customer portal URL
      required:
        - url

    PriceResponse:
      type: object
      properties:
        priceId:
          type: string
          nullable: true
        amount:
          type: number
        currency:
          type: string
      required:
        - amount
        - currency

    SaltResponse:
      type: object
      properties:
        salt:
          type: string
          description: Per-user random hex salt for client-side key derivation
      required:
        - salt

    BlobResponse:
      type: object
      properties:
        blob:
          type: string
          nullable: true
          description: Base64-encoded AES-GCM encrypted payload

    BlobUploadRequest:
      type: object
      properties:
        blob:
          type: string
          description: Base64-encoded AES-GCM encrypted payload
      required:
        - blob

    BlobUploadResponse:
      type: object
      properties:
        ok:
          type: boolean
        updatedAt:
          type: string
      required:
        - ok
        - updatedAt